requirements.txt
opencv-python
face_recognition
python-telegram-bot
PyYAML
numpy